backend/requirements.txt
fastapi>=0.109.2
uvicorn==0.27.1
pydantic>=2.7.1
playwright>=1.43.0
hyperbrowser>=0.1.0
python-dotenv==1.0.1
requests==2.31.0 